bugfix for DFS & Dijkstra
Moved print functions from main.cpp
