+++ app/DFS.h
@@ -12,6 +12,7 @@
 #include <algorithm>
 #include <string>
 #include <QHash>
+#include <iostream>
 
 using namespace std;
 
@@ -22,5 +23,11 @@ using namespace std;
 // for the shortest path.
 //QHash<QString, StadiumDistance*> DFS(const StadiumGraph& graph, Stadium* start);
 QVector<StadiumDistance*> DFS(const StadiumGraph& graph, Stadium* start);
+
+// Creates & prints the specified path for DFS
+void runDFS(const StadiumGraph& graph, QString start);
+
+// Runs / calls path for DFS from start stadium
+void testDFS();
 
 #endif // DFS_H
