+++ app/mainwindow.cpp
@@ -6,15 +6,16 @@
 #include "Database.h"
 #include "Team.h"
 #include "Dijkstra.h"
+#include "DFS.h"
 
 MainWindow::MainWindow(QWidget *parent)
-	: QMainWindow(parent)
-	, ui(new Ui::MainWindow)
+    : QMainWindow(parent)
+    , ui(new Ui::MainWindow)
 {
-	ui->setupUi(this);
+    ui->setupUi(this);
 
     //PROCESSING - Setting Main Window Title
-	this -> setWindowTitle("Let's Play Football");
+    this -> setWindowTitle("Let's Play Football");
 
     /************************************************************
     * PROCESSING - Assign an image into new QPixmap variable
@@ -29,11 +30,14 @@ MainWindow::MainWindow(QWidget *parent)
     ui -> Image_Label -> setPixmap(homePageBackground.scaled
                                   (ImageLabelW, ImageLabelH));
 
+    //TEST DIJKSTRA & DFS
+    testDijkstra();
+    testDFS();
 }
 
 MainWindow::~MainWindow()
 {
-	delete ui;
+    delete ui;
 }
 
 
@@ -44,7 +48,7 @@ void MainWindow::on_Admin_Page_PushButton_clicked()
     Login *loginUi;
     loginUi = new Login(this);
 
-	hide();
+    hide();
 
     loginUi -> show();
 }
@@ -56,7 +60,7 @@ void MainWindow::on_Display_Team_Stadium_Info_clicked()
     DisplayInfo *displayUI;
     displayUI = new DisplayInfo(this);
 
-	hide();
+    hide();
 
     displayUI -> show();
 }
@@ -65,10 +69,10 @@ void MainWindow::on_Display_Team_Stadium_Info_clicked()
 /// hidden before showing the Travel page
 void MainWindow::on_Souvenir_And_Trip_PushButton_clicked()
 {
-	SouvenirAndTrip *ui;
-	ui = new SouvenirAndTrip(this);
+    SouvenirAndTrip *ui;
+    ui = new SouvenirAndTrip(this);
 
-	hide();
+    hide();
 
-	ui -> show();
+    ui -> show();
 }
